The Index update to end all Index updates
This is the full version, all features implemented for login, registration, and forgot password. I decided to not implement image uploading however as it EATS through database storage. It's not practical. Everything else is good to go though, and connections to our actual database should be working. As for all these new files, it's just for email generation, it's supposed to be a folder.